+++ 00-Backend/InternshipPortal/src/main/java/com/subodh/InternshipPortal/modals/MastersEducation.java
@@ -27,5 +27,5 @@ public class MastersEducation {
 
     @DecimalMin(value = "0.00", inclusive = false)
     @DecimalMax(value = "100.00", inclusive = false)
-    private Long passingPercentage;
+    private Double passingPercentage;
 }
